+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.1.2
+
+* added padPartial method
+* decode will automatically pad a partial IAC
+
 ## v0.1.1
 
 * isValid now uses OLC.isFull instead of OLC.isValid for checking Open Location Codes